dtw.js server-side only
This code does not use any client-side variables. It therefore runs more smoothly than the original code.
Furthermore, I got rid of the ee.Algorithms.If functions. The time differences between all time steps are first calculated and saved under the variable 'dt_list'. Time-constrained DTW then uses a filter (dt <= beta) to pass over the time steps that fulfill the condition.
The code is still not perfect. The timestamps of the image time series are not used by the code, but simply the timestamps of the patterns (line 102). With other words, here we make the assumption that the number of timestamps of the image time series is equal to the number of time stamps of the patterns, and that the time steps are equal. However, the timestamps of the image time series are usually available from the image properties ('system:time_start'), so they could be pulled from there.
I have commented out the code for distance type 'angular'. I haven't revised this section yet and it therefore still uses client-side for loops.